fyp-ai/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.10.0=pypi_0
aiohttp=3.6.2=pypi_0
argon2-cffi=20.1.0=pypi_0
astunparse=1.6.3=pypi_0
async-generator=1.10=pypi_0
async-timeout=3.0.1=pypi_0
attrs=20.2.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.9.2=pypi_0
bleach=3.2.1=pypi_0
blis=0.4.1=pypi_0
ca-certificates=2020.7.22=0
cachetools=4.1.1=pypi_0
catalogue=1.0.0=pypi_0
certifi=2020.6.20=py38_0
cffi=1.14.3=pypi_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
cssselect=1.1.0=pypi_0
cycler=0.10.0=pypi_0
cymem=2.0.3=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
en-core-web-lg=2.3.1=pypi_0
entrypoints=0.3=pypi_0
feedfinder2=0.0.4=pypi_0
feedparser=6.0.1=pypi_0
gast=0.3.3=pypi_0
gensim=3.8.3=pypi_0
google-auth=1.22.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.32.0=pypi_0
h5py=2.10.0=pypi_0
idna=2.10=pypi_0
ipykernel=5.3.4=pypi_0
ipython=7.18.1=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.17.2=pypi_0
jieba3k=0.35.1=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.16.0=pypi_0
json5=0.9.5=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-core=4.6.3=pypi_0
jupyterlab=2.2.8=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-server=1.2.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.2.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
lxml=4.5.2=pypi_0
markdown=3.2.2=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.2=pypi_0
mistune=0.8.4=pypi_0
multidict=4.7.6=pypi_0
murmurhash=1.0.2=pypi_0
nbclient=0.5.0=pypi_0
nbconvert=6.0.6=pypi_0
nbformat=5.0.7=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=pypi_0
newspaper3k=0.2.8=pypi_0
nltk=3.5=pypi_0
notebook=6.1.4=pypi_0
numpy=1.18.5=pypi_0
oauthlib=3.1.0=pypi_0
openssl=1.1.1h=h7b6447c_0
opt-einsum=3.3.0=pypi_0
packaging=20.4=pypi_0
pandas=1.1.2=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.7.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.2.0=pypi_0
pip=20.2.2=py38_0
plac=1.1.3=pypi_0
preshed=3.0.2=pypi_0
prometheus-client=0.8.0=pypi_0
prompt-toolkit=3.0.7=pypi_0
protobuf=3.13.0=pypi_0
ptyprocess=0.6.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pygments=2.7.1=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=pypi_0
pytz=2020.1=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=pypi_0
readline=8.0=h7b6447c_0
regex=2020.9.27=pypi_0
requests=2.24.0=pypi_0
requests-file=1.5.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.2=pypi_0
seaborn=0.11.0=pypi_0
send2trash=1.5.0=pypi_0
setuptools=49.6.0=py38_1
sgmllib3k=1.0.0=pypi_0
six=1.15.0=pypi_0
smart-open=2.2.0=pypi_0
soupsieve=2.0.1=pypi_0
spacy=2.3.2=pypi_0
spacy-lookups-data=0.3.2=pypi_0
sqlite=3.33.0=h62c20be_0
srsly=1.0.2=pypi_0
tensorboard=2.3.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorflow=2.3.1=pypi_0
tensorflow-estimator=2.3.0=pypi_0
tensorflow-hub=0.10.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.1=pypi_0
testpath=0.4.4=pypi_0
thinc=7.4.1=pypi_0
threadpoolctl=2.1.0=pypi_0
tinysegmenter=0.3=pypi_0
tk=8.6.10=hbc83047_0
tldextract=2.2.3=pypi_0
tornado=6.0.4=pypi_0
tqdm=4.50.0=pypi_0
traitlets=5.0.4=pypi_0
urllib3=1.25.10=pypi_0
wasabi=0.8.0=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.35.1=py_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.6.0=pypi_0
zlib=1.2.11=h7b6447c_3